pyproject.toml
[tool.poetry]
name = "goods_scrapper"
version = "0.1.0"
description = ""
authors = ["Dector <roman_dector@vk.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fake-useragent = "^0.1.11"
aiohttp = "^3.8.1"
beautifulsoup4 = "^4.11.1"
openpyxl = "^3.0.10"
python-dotenv = "^0.20.0"
aiosqlite = "^0.17.0"
lxml = "^4.9.0"
textual = "^0.1.18"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.961"
bpython = "^0.22.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
